build-it.sh now records Git hash when on clean workspace to avoid rebuilds.

+++ tools/real-clean-workspace.sh
@@ -5,6 +5,16 @@ set -e
 
 cd $(dirname $0)/..
 
+declare CODEKVAST_VERSION=$(grep codekvastVersion gradle.properties | egrep --only-matching '[0-9.]+')
+declare GIT_HASH=$(git rev-parse --short HEAD)
+declare BUILD_STATE_FILE=.buildState
+
+declare lastBuilt=
+if [ "$(cat ${BUILD_STATE_FILE} 2>/dev/null)" == "${CODEKVAST_VERSION}-${GIT_HASH}" ]; then
+  echo "No need to clean, build is up-to-date with ${CODEKVAST_VERSION}-${GIT_HASH}"
+  exit 0
+fi
+
 echo "Cleaning Gradle build state..."
 rm -fr ./.gradle
